Add type validation in params, queries headers
Add type validation in params, queries headers. Extract common functionality.
