+++ lib/ajValidation/ajvValidation.js
@@ -1,23 +1,12 @@
 const { formatDataPath,
   formatSchemaPathFromAJVErrorToConvertToDataPath,
-  typesMap } = require('../common/schemaUtilsCommon');
+  isTypeValue } = require('../common/schemaUtilsCommon');
 
 var _ = require('lodash');
 const IGNORED_KEYWORDS = ['propertyNames', 'const', 'additionalItems', 'dependencies'],
   { validateSchemaAJV } = require('./ajvValidator'),
   { validateSchemaAJVDraft04 } = require('./ajvValidatorDraft04'),
   specialDraft = 'http://json-schema.org/draft-04/schema#';
-
-/**
- * Checks if value is postman variable or not
- *
- * @param {string} type - type to look for
- * @param {string} format - format from schema
- * @returns {Boolean} postman variable or not
- */
-function getDefaultFromTypeAndFormat(type, format) {
-  return typesMap[type][format];
-}
 
 /**
  * Checks if value is postman variable or not
@@ -29,104 +18,6 @@ function isPmVariable (value) {
   // collection/environment variables are in format - {{var}}
   return _.isString(value) && _.startsWith(value, '{{') && _.endsWith(value, '}}');
 }
-
-/**
- * Checks if value is the representation of its type like:
- * "<integer>"
- *
- * @param {string} value - Value to check for
- * @param {string} type - The type in the schemna
- * @returns {Boolean} the value is the representation of its type
- */
-function compareType(value, type) {
-  return value === '<' + type + '>';
-}
-
-/**
- * Checks if value is the representation of its type like:
- * "<integer>"
- * Works in array types
- * @param {string} value - Value to check for
- * @param {*} types - The types in the schemna
- * @returns {Boolean} the value is the representation of its type
- */
-function isTypeValueArrayCheck(value, types) {
-  return types.find((type) => {
-    return compareType(value, type);
-  }) !== undefined;
-}
-
-/**
- * Checks if value is the representation of its type like:
- * "<integer>"
- * Works in array types
- * @param {string} value - Value to check for
- * @param {*} types - The types in the schemna
- * @returns {Boolean} the value is the representation of its type
- */
-function checkValueOnlyTypes(value, types) {
-  return Array.isArray(types) ? isTypeValueArrayCheck(value, types) : compareType(value, types);
-}
-
-/**
- * Checks if value is the representation of its type like:
- * "<integer>"
- * Works in array types
- * @param {string} value - Value to check for
- * @param {*} types - The types in the schema
- * @param {*} format - format from the schema
- * @returns {Boolean} the value is the representation of its type
- */
-function checkValueTypesAndFormat(value, types, format) {
-  let typesNotInMapp = [],
-    typesArray = Array.isArray(types) ? types : [types],
-    found = typesArray.find((type) => {
-      let defaultValue;
-      if (typesMap.hasOwnProperty(type)) {
-        defaultValue = getDefaultFromTypeAndFormat(type, format);
-
-        // in case the format is a custom format (email, hostname etc.)
-        // https://swagger.io/docs/specification/data-models/data-types/#string
-        if (!defaultValue && format) {
-          defaultValue = '<' + format + '>';
-        }
-      }
-      else {
-        typesNotInMapp.push(type);
-      }
-      return defaultValue === value;
-    });
-
-  if (found) {
-    return true;
-  }
-
-  found = typesNotInMapp.find((type) => {
-    let defaultValue;
-    defaultValue = '<' + type + (format ? ('-' + format) : '') + '>';
-    return defaultValue === value;
-  });
-
-  return found !== undefined;
-}
-
-/**
- * Checks if value is the representation of its type like:
- * "<integer>"
- * Works in array types
- * @param {string} value - Value to check for
- * @param {*} schema - The schema portion used in validation
- * @returns {Boolean} the value is the representation of its type
- */
-function isTypeValue(value, schema) {
-  if (schema.hasOwnProperty('type') && !schema.hasOwnProperty('format')) {
-    return checkValueOnlyTypes(value, schema.type);
-  }
-  if (schema.hasOwnProperty('type') && schema.hasOwnProperty('format')) {
-    return checkValueTypesAndFormat(value, schema.type, schema.format);
-  }
-}
-
 
 /**
  * returns the local $schema value
