+++ lib/common/schemaUtilsCommon.js
@@ -3,6 +3,7 @@
  */
 
 const parse = require('../parse.js'),
+  _ = require('lodash'),
   typesMap = {
     integer: {
       int32: '<integer>',
@@ -20,6 +21,16 @@ const parse = require('../parse.js'),
       password: '<password>'
     },
     boolean: '<boolean>'
+  },
+
+  /* eslint-disable arrow-body-style */
+  schemaTypeToJsValidator = {
+    'string': (d) => typeof d === 'string',
+    'number': (d) => !isNaN(d),
+    'integer': (d) => !isNaN(d) && Number.isInteger(Number(d)),
+    'boolean': (d) => _.isBoolean(d) || d === 'true' || d === 'false',
+    'array': (d) => Array.isArray(d),
+    'object': (d) => typeof d === 'object' && !Array.isArray(d)
   };
 
 /**
@@ -42,6 +53,152 @@ function removeTypeFromLastPosition(schemaPath) {
     splittedDataPath.splice(-1);
   }
   return splittedDataPath.join('/');
+}
+
+/**
+ * Checks if value is the representation of its type like:
+ * "<integer>"
+ *
+ * @param {string} value - Value to check for
+ * @param {string} type - The type in the schemna
+ * @returns {Boolean} the value is the representation of its type
+ */
+function compareType(value, type) {
+  return value === '<' + type + '>';
+}
+
+/**
+ * Checks if value is the representation of its type like:
+ * "<integer>"
+ * Works in array types
+ * @param {string} value - Value to check for
+ * @param {*} types - The types in the schemna
+ * @returns {Boolean} the value is the representation of its type
+ */
+function isTypeValueArrayCheck(value, types) {
+  return types.find((type) => {
+    return compareType(value, type);
+  }) !== undefined;
+}
+
+/**
+ * Checks if value is the representation of its type like:
+ * "<integer>"
+ * Works in array types
+ * @param {string} value - Value to check for
+ * @param {*} types - The types in the schemna
+ * @returns {Boolean} the value is the representation of its type
+ */
+function checkValueOnlyTypes(value, types) {
+  return Array.isArray(types) ? isTypeValueArrayCheck(value, types) : compareType(value, types);
+}
+
+/**
+ * Checks if value is postman variable or not
+ *
+ * @param {string} type - type to look for
+ * @param {string} format - format from schema
+ * @returns {Boolean} postman variable or not
+ */
+function getDefaultFromTypeAndFormat(type, format) {
+  return typesMap[type][format];
+}
+
+
+/**
+ * Checks if value is the representation of its type like:
+ * "<integer>"
+ * Works in array types
+ * @param {string} value - Value to check for
+ * @param {*} types - The types in the schema
+ * @param {*} format - format from the schema
+ * @returns {Boolean} the value is the representation of its type
+ */
+function checkValueTypesAndFormat(value, types, format) {
+  let typesNotInMapp = [],
+    typesArray = Array.isArray(types) ? types : [types],
+    found = typesArray.find((type) => {
+      let defaultValue;
+      if (typesMap.hasOwnProperty(type)) {
+        defaultValue = getDefaultFromTypeAndFormat(type, format);
+
+        // in case the format is a custom format (email, hostname etc.)
+        // https://swagger.io/docs/specification/data-models/data-types/#string
+        if (!defaultValue && format) {
+          defaultValue = '<' + format + '>';
+        }
+      }
+      else {
+        typesNotInMapp.push(type);
+      }
+      return defaultValue === value;
+    });
+
+  if (found) {
+    return true;
+  }
+
+  found = typesNotInMapp.find((type) => {
+    let defaultValue;
+    defaultValue = '<' + type + (format ? ('-' + format) : '') + '>';
+    return defaultValue === value;
+  });
+
+  return found !== undefined;
+}
+
+/**
+ * Checks if value is equal to the defined default
+ * Works in array types
+ * @param {string} value - Value to check for
+ * @param {*} definedDefault - The defined default value of the schema
+ * @returns {Boolean} wheter value is equal to the defined or not
+ */
+function checkValueEqualsDefault(value, definedDefault) {
+  return value === definedDefault;
+}
+
+/**
+ * Checks if value is the representation of its type like:
+ * "<integer>"
+ * Works in array types
+ * @param {string} value - Value to check for
+ * @param {*} schema - The schema portion used in validation
+ * @returns {Boolean} the value is the representation of its type
+ */
+function isTypeValue(value, schema) {
+  if (schema.hasOwnProperty('type') && schema.hasOwnProperty('default')) {
+    const isDefault = checkValueEqualsDefault(value, schema.default);
+    if (isDefault) {
+      return true;
+    }
+  }
+  if (schema.hasOwnProperty('type') && !schema.hasOwnProperty('format')) {
+    return checkValueOnlyTypes(value, schema.type);
+  }
+  if (schema.hasOwnProperty('type') && schema.hasOwnProperty('format')) {
+    return checkValueTypesAndFormat(value, schema.type, schema.format);
+  }
+}
+
+/**
+ * Checks if value is correcta according to schema
+ * If the value should be numeric, it tries to convert and then validate
+ * also validates if the value is a correct representation in the form of
+ * <long> or <integer> etc for integers format 32 or format 64
+ * @param {string} value - Value to check for
+ * @param {*} schema - The schema portion used in validation
+ * @returns {Boolean} the value is the representation of its type
+ */
+function checkIsCorrectType(value, schema) {
+  if (schema.hasOwnProperty('type') &&
+    typeof schemaTypeToJsValidator[schema.type] === 'function') {
+    const isCorrectType = schemaTypeToJsValidator[schema.type](value);
+    if (isCorrectType) {
+      return true;
+    }
+  }
+  return isTypeValue(value, schema);
 }
 
 module.exports = {
@@ -158,5 +315,31 @@ module.exports = {
     return removeTypeFromLastPosition(removeSharpAndSlashFromFirstPosition(schemaPath));
   },
 
-  typesMap
+  typesMap,
+
+  /**
+   * Checks if value is the representation of its type like:
+   * "<integer>"
+   * Works in array types
+   * @param {string} value - Value to check for
+   * @param {*} schema - The schema portion used in validation
+   * @returns {Boolean} the value is the representation of its type
+   */
+  isTypeValue,
+
+  /**
+   * Checks if value is correcta according to schema
+   * If the value should be numeric, it tries to convert and then validate
+   * also validates if the value is a correct representation in the form of
+   * <long> or <integer> etc for integers format 32 or format 64
+   * @param {string} value - Value to check for
+   * @param {*} schema - The schema portion used in validation
+   * @returns {Boolean} the value is the representation of its type
+   */
+  checkIsCorrectType,
+
+
+  isKnownType: function(schema) {
+    return typeof schemaTypeToJsValidator[schema.type] === 'function';
+  }
 };
