+++ test/unit/schemaUtilsCommon.test.js
@@ -1,5 +1,7 @@
 const schemaUtilsCommon = require('../../lib/common/schemaUtilsCommon'),
-  { formatDataPath, formatSchemaPathFromAJVErrorToConvertToDataPath } = require('../../lib/common/schemaUtilsCommon'),
+  { formatDataPath,
+    formatSchemaPathFromAJVErrorToConvertToDataPath,
+    isTypeValue } = require('../../lib/common/schemaUtilsCommon'),
   expect = require('chai').expect;
 
 describe('formatData method', function() {
@@ -164,4 +166,101 @@ describe('formatSchemaPathFromAJVErrorToConvertToDataPath method', function () {
       formatSchemaPathFromAJVErrorToConvertToDataPath('#/properties/automatic/items/properties/configs/items/type');
     expect(result).to.equal('properties/automatic/items/properties/configs/items');
   });
+});
+
+describe('isTypeValue method', function () {
+  it('should return true when value is <integer> and type is integer', function () {
+    const result = isTypeValue('<integer>', {
+      type: [
+        'integer'
+      ]
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return true when input is <long> type integer and format int64', function () {
+    const result = isTypeValue('<long>', {
+      format: 'int64',
+      type: ['integer']
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return true when value is <uuid> type is string format is uuid', function () {
+    const result = isTypeValue('<uuid>', {
+      format: 'uuid',
+      type: ['string']
+    });
+    expect(result).to.be.true;
+  });
+
+
+  it('should return true when value is <otherType> type is otherType and there is not format', function () {
+    const result = isTypeValue('<otherType>', {
+      type: ['otherType']
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return true value is <otherType-otherFormat> type is otherType and format is otherFormat', function () {
+    const result = isTypeValue('<otherType-otherFormat>', {
+      format: 'otherFormat',
+      type: ['otherType']
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return false when value is <integer> and type is boolean', function () {
+    const result = isTypeValue('<integer>', {
+      type: ['boolean']
+    });
+    expect(result).to.be.false;
+  });
+
+  it('should return true when value is <string> and type is string', function () {
+    const result = isTypeValue('<string>', {
+      type: ['string']
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return true when value is <integer> and type is ["boolean", "integer"]', function () {
+    const result = isTypeValue('<integer>', {
+      type: ['boolean', 'integer']
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return true when value is <integer> and type is integer not array', function () {
+    const result = isTypeValue('<integer>', {
+      type: 'integer'
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return true when value is <integer> and type is integer not array format int64', function () {
+    const result = isTypeValue('<long>', {
+      format: 'int64',
+      type: 'integer'
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return true when value is <long> and type is integer, format not present default is <long>', function () {
+    const result = isTypeValue('<long>', {
+      default: '<long>',
+      type: 'integer'
+    });
+    expect(result).to.be.true;
+  });
+
+  it('should return false when value is <long> and type is integer,' +
+    ' there is no format default is <nlong>', function () {
+    const result = isTypeValue('<nlong>', {
+      default: '<long>',
+      type: 'integer'
+    });
+    expect(result).to.be.false;
+  });
+
 });
